Refa: abstract figure parser as a separate module
